feat(xml): move function for xml parse to a specific file
